Fix build on Windows
Windows has a different signature for the syscall.CloseOnExec func and
based on quick research Windows's CloseOnExec might also behave
differently. [1, 2]

This commit moves the socket activation code into a
conditionally-compiled file, with no-op implementations for Windows. The
current implementation ignores the socket activation environment
variables.

[1] https://github.com/cloudflare/tableflip/issues/40
[2] https://docs.microsoft.com/en-us/windows/win32/sysinfo/handles-and-objects
